backend/conditionchecking_microservice/requirements.txt
Flask==2.1.3
Werkzeug==2.1.2
firebase-admin==6.0.0
requests==2.28.1
python-dotenv==0.21.1

